Removed reflection and other extraneous things from the core library. (#158)
* Removed reflection and other extraneous things from the core library.

* Added missing files and ran buildifier.

* New CMakeLists.txt.

* Made table its own cc_library() for internal usage.
