+++ upb/json/parser.c
@@ -1533,7 +1533,6 @@ static bool end_duration_base(upb_json_parser *p, const char *ptr) {
 static int parse_timestamp_number(upb_json_parser *p) {
   size_t len;
   const char *buf;
-  char *end;
   int val;
 
   /* atoi() and friends unfortunately do not support specifying the length of
@@ -2009,7 +2008,6 @@ static void end_any_member(upb_json_parser *p, const char *ptr) {
 
 static bool start_subobject(upb_json_parser *p) {
   if (p->top->is_unknown_field) {
-    upb_jsonparser_frame *inner;
     if (!check_stack(p)) return false;
 
     p->top = start_jsonparser_frame(p);
@@ -2544,11 +2542,11 @@ static bool does_fieldmask_end(upb_json_parser *p) {
  * final state once, when the closing '"' is seen. */
 
 
-#line 2749 "upb/json/parser.rl"
+#line 2747 "upb/json/parser.rl"
 
 
 
-#line 2552 "upb/json/parser.c"
+#line 2550 "upb/json/parser.c"
 static const char _json_actions[] = {
 	0, 1, 0, 1, 1, 1, 3, 1, 
 	4, 1, 6, 1, 7, 1, 8, 1, 
@@ -2803,7 +2801,7 @@ static const int json_en_value_machine = 78;
 static const int json_en_main = 1;
 
 
-#line 2752 "upb/json/parser.rl"
+#line 2750 "upb/json/parser.rl"
 
 size_t parse(void *closure, const void *hd, const char *buf, size_t size,
              const upb_bufhandle *handle) {
@@ -2826,7 +2824,7 @@ size_t parse(void *closure, const void *hd, const char *buf, size_t size,
   capture_resume(parser, buf);
 
   
-#line 2830 "upb/json/parser.c"
+#line 2828 "upb/json/parser.c"
 	{
 	int _klen;
 	unsigned int _trans;
@@ -2901,147 +2899,147 @@ _match:
 		switch ( *_acts++ )
 		{
 	case 1:
-#line 2557 "upb/json/parser.rl"
+#line 2555 "upb/json/parser.rl"
 	{ p--; {cs = stack[--top]; goto _again;} }
 	break;
 	case 2:
-#line 2559 "upb/json/parser.rl"
+#line 2557 "upb/json/parser.rl"
 	{ p--; {stack[top++] = cs; cs = 23;goto _again;} }
 	break;
 	case 3:
-#line 2563 "upb/json/parser.rl"
+#line 2561 "upb/json/parser.rl"
 	{ start_text(parser, p); }
 	break;
 	case 4:
-#line 2564 "upb/json/parser.rl"
+#line 2562 "upb/json/parser.rl"
 	{ CHECK_RETURN_TOP(end_text(parser, p)); }
 	break;
 	case 5:
-#line 2570 "upb/json/parser.rl"
+#line 2568 "upb/json/parser.rl"
 	{ start_hex(parser); }
 	break;
 	case 6:
-#line 2571 "upb/json/parser.rl"
+#line 2569 "upb/json/parser.rl"
 	{ hexdigit(parser, p); }
 	break;
 	case 7:
-#line 2572 "upb/json/parser.rl"
+#line 2570 "upb/json/parser.rl"
 	{ CHECK_RETURN_TOP(end_hex(parser)); }
 	break;
 	case 8:
-#line 2578 "upb/json/parser.rl"
+#line 2576 "upb/json/parser.rl"
 	{ CHECK_RETURN_TOP(escape(parser, p)); }
 	break;
 	case 9:
-#line 2584 "upb/json/parser.rl"
+#line 2582 "upb/json/parser.rl"
 	{ p--; {cs = stack[--top]; goto _again;} }
 	break;
 	case 10:
-#line 2589 "upb/json/parser.rl"
+#line 2587 "upb/json/parser.rl"
 	{ start_year(parser, p); }
 	break;
 	case 11:
-#line 2590 "upb/json/parser.rl"
+#line 2588 "upb/json/parser.rl"
 	{ CHECK_RETURN_TOP(end_year(parser, p)); }
 	break;
 	case 12:
-#line 2594 "upb/json/parser.rl"
+#line 2592 "upb/json/parser.rl"
 	{ start_month(parser, p); }
 	break;
 	case 13:
-#line 2595 "upb/json/parser.rl"
+#line 2593 "upb/json/parser.rl"
 	{ CHECK_RETURN_TOP(end_month(parser, p)); }
 	break;
 	case 14:
-#line 2599 "upb/json/parser.rl"
+#line 2597 "upb/json/parser.rl"
 	{ start_day(parser, p); }
 	break;
 	case 15:
-#line 2600 "upb/json/parser.rl"
+#line 2598 "upb/json/parser.rl"
 	{ CHECK_RETURN_TOP(end_day(parser, p)); }
 	break;
 	case 16:
-#line 2604 "upb/json/parser.rl"
+#line 2602 "upb/json/parser.rl"
 	{ start_hour(parser, p); }
 	break;
 	case 17:
-#line 2605 "upb/json/parser.rl"
+#line 2603 "upb/json/parser.rl"
 	{ CHECK_RETURN_TOP(end_hour(parser, p)); }
 	break;
 	case 18:
-#line 2609 "upb/json/parser.rl"
+#line 2607 "upb/json/parser.rl"
 	{ start_minute(parser, p); }
 	break;
 	case 19:
-#line 2610 "upb/json/parser.rl"
+#line 2608 "upb/json/parser.rl"
 	{ CHECK_RETURN_TOP(end_minute(parser, p)); }
 	break;
 	case 20:
-#line 2614 "upb/json/parser.rl"
+#line 2612 "upb/json/parser.rl"
 	{ start_second(parser, p); }
 	break;
 	case 21:
-#line 2615 "upb/json/parser.rl"
+#line 2613 "upb/json/parser.rl"
 	{ CHECK_RETURN_TOP(end_second(parser, p)); }
 	break;
 	case 22:
-#line 2620 "upb/json/parser.rl"
+#line 2618 "upb/json/parser.rl"
 	{ start_duration_base(parser, p); }
 	break;
 	case 23:
-#line 2621 "upb/json/parser.rl"
+#line 2619 "upb/json/parser.rl"
 	{ CHECK_RETURN_TOP(end_duration_base(parser, p)); }
 	break;
 	case 24:
-#line 2623 "upb/json/parser.rl"
+#line 2621 "upb/json/parser.rl"
 	{ p--; {cs = stack[--top]; goto _again;} }
 	break;
 	case 25:
-#line 2628 "upb/json/parser.rl"
+#line 2626 "upb/json/parser.rl"
 	{ start_timestamp_base(parser); }
 	break;
 	case 26:
-#line 2630 "upb/json/parser.rl"
+#line 2628 "upb/json/parser.rl"
 	{ start_timestamp_fraction(parser, p); }
 	break;
 	case 27:
-#line 2631 "upb/json/parser.rl"
+#line 2629 "upb/json/parser.rl"
 	{ CHECK_RETURN_TOP(end_timestamp_fraction(parser, p)); }
 	break;
 	case 28:
-#line 2633 "upb/json/parser.rl"
+#line 2631 "upb/json/parser.rl"
 	{ start_timestamp_zone(parser, p); }
 	break;
 	case 29:
-#line 2634 "upb/json/parser.rl"
+#line 2632 "upb/json/parser.rl"
 	{ CHECK_RETURN_TOP(end_timestamp_zone(parser, p)); }
 	break;
 	case 30:
-#line 2636 "upb/json/parser.rl"
+#line 2634 "upb/json/parser.rl"
 	{ p--; {cs = stack[--top]; goto _again;} }
 	break;
 	case 31:
-#line 2641 "upb/json/parser.rl"
+#line 2639 "upb/json/parser.rl"
 	{ start_fieldmask_path_text(parser, p); }
 	break;
 	case 32:
-#line 2642 "upb/json/parser.rl"
+#line 2640 "upb/json/parser.rl"
 	{ end_fieldmask_path_text(parser, p); }
 	break;
 	case 33:
-#line 2647 "upb/json/parser.rl"
+#line 2645 "upb/json/parser.rl"
 	{ start_fieldmask_path(parser); }
 	break;
 	case 34:
-#line 2648 "upb/json/parser.rl"
+#line 2646 "upb/json/parser.rl"
 	{ end_fieldmask_path(parser); }
 	break;
 	case 35:
-#line 2654 "upb/json/parser.rl"
+#line 2652 "upb/json/parser.rl"
 	{ p--; {cs = stack[--top]; goto _again;} }
 	break;
 	case 36:
-#line 2659 "upb/json/parser.rl"
+#line 2657 "upb/json/parser.rl"
 	{
         if (is_wellknown_msg(parser, UPB_WELLKNOWN_TIMESTAMP)) {
           {stack[top++] = cs; cs = 47;goto _again;}
@@ -3055,11 +3053,11 @@ _match:
       }
 	break;
 	case 37:
-#line 2672 "upb/json/parser.rl"
+#line 2670 "upb/json/parser.rl"
 	{ p--; {stack[top++] = cs; cs = 78;goto _again;} }
 	break;
 	case 38:
-#line 2677 "upb/json/parser.rl"
+#line 2675 "upb/json/parser.rl"
 	{
         if (is_wellknown_msg(parser, UPB_WELLKNOWN_ANY)) {
           start_any_member(parser, p);
@@ -3069,11 +3067,11 @@ _match:
       }
 	break;
 	case 39:
-#line 2684 "upb/json/parser.rl"
+#line 2682 "upb/json/parser.rl"
 	{ CHECK_RETURN_TOP(end_membername(parser)); }
 	break;
 	case 40:
-#line 2687 "upb/json/parser.rl"
+#line 2685 "upb/json/parser.rl"
 	{
         if (is_wellknown_msg(parser, UPB_WELLKNOWN_ANY)) {
           end_any_member(parser, p);
@@ -3083,7 +3081,7 @@ _match:
       }
 	break;
 	case 41:
-#line 2698 "upb/json/parser.rl"
+#line 2696 "upb/json/parser.rl"
 	{
         if (is_wellknown_msg(parser, UPB_WELLKNOWN_ANY)) {
           start_any_object(parser, p);
@@ -3093,7 +3091,7 @@ _match:
       }
 	break;
 	case 42:
-#line 2707 "upb/json/parser.rl"
+#line 2705 "upb/json/parser.rl"
 	{
         if (is_wellknown_msg(parser, UPB_WELLKNOWN_ANY)) {
           CHECK_RETURN_TOP(end_any_object(parser, p));
@@ -3103,54 +3101,54 @@ _match:
       }
 	break;
 	case 43:
-#line 2719 "upb/json/parser.rl"
+#line 2717 "upb/json/parser.rl"
 	{ CHECK_RETURN_TOP(start_array(parser)); }
 	break;
 	case 44:
-#line 2723 "upb/json/parser.rl"
+#line 2721 "upb/json/parser.rl"
 	{ end_array(parser); }
 	break;
 	case 45:
-#line 2728 "upb/json/parser.rl"
+#line 2726 "upb/json/parser.rl"
 	{ CHECK_RETURN_TOP(start_number(parser, p)); }
 	break;
 	case 46:
-#line 2729 "upb/json/parser.rl"
+#line 2727 "upb/json/parser.rl"
 	{ CHECK_RETURN_TOP(end_number(parser, p)); }
 	break;
 	case 47:
-#line 2731 "upb/json/parser.rl"
+#line 2729 "upb/json/parser.rl"
 	{ CHECK_RETURN_TOP(start_stringval(parser)); }
 	break;
 	case 48:
-#line 2732 "upb/json/parser.rl"
+#line 2730 "upb/json/parser.rl"
 	{ CHECK_RETURN_TOP(end_stringval(parser)); }
 	break;
 	case 49:
-#line 2734 "upb/json/parser.rl"
+#line 2732 "upb/json/parser.rl"
 	{ CHECK_RETURN_TOP(end_bool(parser, true)); }
 	break;
 	case 50:
-#line 2736 "upb/json/parser.rl"
+#line 2734 "upb/json/parser.rl"
 	{ CHECK_RETURN_TOP(end_bool(parser, false)); }
 	break;
 	case 51:
-#line 2738 "upb/json/parser.rl"
+#line 2736 "upb/json/parser.rl"
 	{ CHECK_RETURN_TOP(end_null(parser)); }
 	break;
 	case 52:
-#line 2740 "upb/json/parser.rl"
+#line 2738 "upb/json/parser.rl"
 	{ CHECK_RETURN_TOP(start_subobject_full(parser)); }
 	break;
 	case 53:
-#line 2741 "upb/json/parser.rl"
+#line 2739 "upb/json/parser.rl"
 	{ end_subobject_full(parser); }
 	break;
 	case 54:
-#line 2746 "upb/json/parser.rl"
+#line 2744 "upb/json/parser.rl"
 	{ p--; {cs = stack[--top]; goto _again;} }
 	break;
-#line 3154 "upb/json/parser.c"
+#line 3152 "upb/json/parser.c"
 		}
 	}
 
@@ -3167,32 +3165,32 @@ _again:
 	while ( __nacts-- > 0 ) {
 		switch ( *__acts++ ) {
 	case 0:
-#line 2555 "upb/json/parser.rl"
+#line 2553 "upb/json/parser.rl"
 	{ p--; {cs = stack[--top]; 	if ( p == pe )
 		goto _test_eof;
 goto _again;} }
 	break;
 	case 46:
-#line 2729 "upb/json/parser.rl"
+#line 2727 "upb/json/parser.rl"
 	{ CHECK_RETURN_TOP(end_number(parser, p)); }
 	break;
 	case 49:
-#line 2734 "upb/json/parser.rl"
+#line 2732 "upb/json/parser.rl"
 	{ CHECK_RETURN_TOP(end_bool(parser, true)); }
 	break;
 	case 50:
-#line 2736 "upb/json/parser.rl"
+#line 2734 "upb/json/parser.rl"
 	{ CHECK_RETURN_TOP(end_bool(parser, false)); }
 	break;
 	case 51:
-#line 2738 "upb/json/parser.rl"
+#line 2736 "upb/json/parser.rl"
 	{ CHECK_RETURN_TOP(end_null(parser)); }
 	break;
 	case 53:
-#line 2741 "upb/json/parser.rl"
+#line 2739 "upb/json/parser.rl"
 	{ end_subobject_full(parser); }
 	break;
-#line 3196 "upb/json/parser.c"
+#line 3194 "upb/json/parser.c"
 		}
 	}
 	}
@@ -3200,7 +3198,7 @@ goto _again;} }
 	_out: {}
 	}
 
-#line 2774 "upb/json/parser.rl"
+#line 2772 "upb/json/parser.rl"
 
   if (p != pe) {
     upb_status_seterrf(parser->status, "Parse error at '%.*s'\n", pe - p, p);
@@ -3243,13 +3241,13 @@ static void json_parser_reset(upb_json_parser *p) {
 
   /* Emit Ragel initialization of the parser. */
   
-#line 3247 "upb/json/parser.c"
+#line 3245 "upb/json/parser.c"
 	{
 	cs = json_start;
 	top = 0;
 	}
 
-#line 2816 "upb/json/parser.rl"
+#line 2814 "upb/json/parser.rl"
   p->current_state = cs;
   p->parser_top = top;
   accumulate_clear(p);
